requirements.txt
flask
openai
transformers
pandas
tabulate
jupyter
python-dotenv